+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.0
 
 - **Breaking**: Replace both `requestAndConfirmAirdropIfRequired()` and `requestAndConfirmAirdrop()` with a single function, `airdropIfRequired()`. See [README.md]!
+- Fix error handling in `confirmTransaction()` to throw errors correctly.
 
 ## 1.5